replaced backlog databases with per-instance tasks
The use of databases to store the backlog of projects and ballots
to denormalize was not reliable when using the high replication datastore.
It has been replaced with scheduling a task for each ballot and project
that does the denormalization.

The util app has been split in to util and system. This allows project and
ballot to use the services from util, without being mixed up with the
high level features in system (such as CSV import) that needs to use
the project, ballot and timeline apps.