add support for JSON import and export
Added JSON as an import and export database format. It assumes
a simple model where there is a JSON attribute for each field in
the model. The only extra processing is the ability to use a
task group to specifiy a project in a ballot, rather than using
project_id, which requires knowing the project's primary key.